Improved resolution/quality of camouflage masks + enabled camo paint on chassis wheels
A few details on the hull (mainly the schurzen plating) looked a bit
pixelated with camouflage paint applied - even on high settings - so, I
made a new camo mask for the hull, and upped the resolution of the rest
of the camo textures. This should make the tank look a bit better
overall (at least, IMO).